Centralize RunAI streamer tensor ownership helper
